pyproject.toml
[project]
name = "ocr-captcha-v3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.1.1",
    "torch>=2.10.0",
    "transformers>=5.1.0",
]

[dependency-groups]
dev = [
    "poethepoet>=0.41.0",
    "ruff>=0.15.1",
]


[tool.poe.tasks]
download = "python -c 'import pretrained'"
typecheck = "ty check"
lint = ["ruff:check", "typecheck"]
fix = ["ruff:format", "ruff:fix", "typecheck"]

[tool.poe.tasks."ruff:format"]
cmd = "ruff format"

[tool.poe.tasks."ruff:check"]
cmd = "ruff check"

[tool.poe.tasks."ruff:fix"]
cmd = "ruff check --fix"

[tool.poe]
executor = { type = "uv" }